chore: apk compatiable

+++ luci-app-openclash/luasrc/controller/openclash.lua
@@ -104,6 +104,8 @@ local device_arh = luci.sys.exec("uname -m |tr -d '\n'")
 
 if pcall(require, "luci.model.ipkg") then
 	opkg = require "luci.model.ipkg"
+else
+	opkg = nil
 end
 
 local core_path_mode = uci:get("openclash", "config", "small_flash_memory")
@@ -216,7 +218,11 @@ local function coremodel()
 	if opkg and opkg.info("libc") and opkg.info("libc")["libc"] then
 		return opkg.info("libc")["libc"]["Architecture"]
 	else
-		return luci.sys.exec("rm -f /var/lock/opkg.lock && opkg status libc 2>/dev/null |grep 'Architecture' |awk -F ': ' '{print $2}' 2>/dev/null")
+		if fs.access("/bin/opkg") then
+			return luci.sys.exec("rm -f /var/lock/opkg.lock && opkg status libc 2>/dev/null |grep 'Architecture' |awk -F ': ' '{print $2}' 2>/dev/null")
+		elseif fs.access("/usr/bin/apk") then
+			return luci.sys.exec("apk list libc |awk '{print $2}'")
+		end
 	end
 end
 
@@ -246,7 +252,11 @@ local function opcv()
 	if opkg and opkg.info("luci-app-openclash") and opkg.info("luci-app-openclash")["luci-app-openclash"] then
 		return "v" .. opkg.info("luci-app-openclash")["luci-app-openclash"]["Version"]
 	else
-		return luci.sys.exec("rm -f /var/lock/opkg.lock && opkg status luci-app-openclash 2>/dev/null |grep 'Version' |awk -F 'Version: ' '{print \"v\"$2}'")
+		if fs.access("/bin/opkg") then
+			return luci.sys.exec("rm -f /var/lock/opkg.lock && opkg status luci-app-openclash 2>/dev/null |grep 'Version' |awk -F 'Version: ' '{print \"v\"$2}'")
+		elseif fs.access("/usr/bin/apk") then
+			return "v" .. luci.sys.exec("apk list luci-app-openclash |grep 'installed' | grep -oE '\\d+(\\.\\d+)*' | head -1")
+		end
 	end
 end
